Crypto.Store.CMS

Description

Cryptographic Message Syntax

RFC 5652: Cryptographic Message Syntax (CMS)

RFC 3370: Cryptographic Message Syntax (CMS) Algorithms

RFC 3560: Use of the RSAES-OAEP Key Transport Algorithm in the Cryptographic Message Syntax (CMS)

RFC 4056: Use of the RSASSA-PSS Signature Algorithm in Cryptographic Message Syntax (CMS)

RFC 3565: Use of the Advanced Encryption Standard (AES) Encryption Algorithm in Cryptographic Message Syntax (CMS)

RFC 5753: Use of Elliptic Curve Cryptography (ECC) Algorithms in Cryptographic Message Syntax (CMS)

RFC 5754: Using SHA2 Algorithms with Cryptographic Message Syntax

RFC 3211: Password-based Encryption for CMS

RFC 5083: Cryptographic Message Syntax (CMS) Authenticated-Enveloped-Data Content Type

RFC 5084: Using AES-CCM and AES-GCM Authenticated Encryption in the Cryptographic Message Syntax (CMS)

RFC 6476: Using Message Authentication Code (MAC) Encryption in the Cryptographic Message Syntax (CMS)

RFC 8103: Using ChaCha20-Poly1305 Authenticated Encryption in the Cryptographic Message Syntax (CMS)

RFC 8418: Use of the Elliptic Curve Diffie-Hellman Key Agreement Algorithm with X25519 and X448 in the Cryptographic Message Syntax (CMS)

RFC 8419: Use of Edwards-Curve Digital Signature Algorithm (EdDSA) Signatures in the Cryptographic Message Syntax (CMS)

RFC 8702: Use of the SHAKE One-Way Hash Functions in the Cryptographic Message Syntax (CMS)

Documentation

data ContentType

CMS content information type.

Constructors

DataType
Arbitrary octet string

SignedDataType
Signed content info

EnvelopedDataType
Enveloped content info

DigestedDataType
Content info with associated digest

EncryptedDataType
Encrypted content info

AuthenticatedDataType
Authenticated content info

AuthEnvelopedDataType
Authenticated-enveloped content info

data ContentInfo

CMS content information.

Constructors

DataCI ByteString
Arbitrary octet string

SignedDataCI (SignedData (Encap EncapsulatedContent))
Signed content info

EnvelopedDataCI (EnvelopedData (Encap EncryptedContent))
Enveloped content info

DigestedDataCI (DigestedData (Encap EncapsulatedContent))
Content info with associated digest

EncryptedDataCI (EncryptedData (Encap EncryptedContent))
Encrypted content info

AuthenticatedDataCI (AuthenticatedData (Encap EncapsulatedContent))
Authenticatedcontent info

AuthEnvelopedDataCI (AuthEnvelopedData (Encap EncryptedContent))
Authenticated-enveloped content info

getContentType :: ContentInfo -> ContentType

Get the type of a content info.

Reading and writing PEM files

readCMSFile :: FilePath -> IO [ContentInfo]

Read content info elements from a PEM file.

readCMSFileFromMemory :: ByteString -> [ContentInfo]

Read content info elements from a bytearray in PEM format.

berToContentInfo :: ByteString -> Either StoreError ContentInfo

Read a content info from a bytearray in BER format.

pemToContentInfo :: [Maybe ContentInfo] -> PEM -> [Maybe ContentInfo]

Read a content info from a PEM element and add it to the accumulator list.

writeCMSFile :: FilePath -> [ContentInfo] -> IO ()

Write content info elements to a PEM file.

writeCMSFileToMemory :: [ContentInfo] -> ByteString

Write content info elements to a bytearray in PEM format.

contentInfoToDER :: ContentInfo -> ByteString

Generate a bytearray in DER format for a content info.

contentInfoToPEM :: ContentInfo -> PEM

Generate PEM for a content info.

Content encapsulation

data Encap a

Denote the state of encapsulated content in a CMS data structure. This type is isomorphic to Maybe.

Constructors

Detached
Content is stored externally to the structure

Attached a
Content is stored inside the CMS struture

fromEncap :: b -> (a -> b) -> Encap a -> b

Fold over an Encap value. This is similar to function maybe. If the content is detached, the first argument is returned. Otherwise the second argument is applied to the content.

class Encapsulates struct

Class of data structures with inner content that may be stored externally. This class has instances for each CMS content type containing other encapsulated or encrypted content info.

Functions fromAttached and fromDetached are used to introspect encapsulation state (attached or detached), and recover a data structure with actionable content.

Functions toAttachedCI and toDetachedCI are needed to decide about the outer encapsulation state and build a ContentInfo.

Minimal complete definition

lens, toCI

isAttached :: Encapsulates struct => struct (Encap a) -> Bool

Return True when the encapsulated content is attached.

fromAttached :: (MonadFail m, Encapsulates struct) => struct (Encap a) -> m (struct a)

Unwrap the encapsulation, assuming the inner content is inside the data structure. The monadic computation fails if the content was detached.

toAttachedCI :: Encapsulates struct => struct ByteString -> ContentInfo

Transform the data structure into a content info, keeping the encapsulated content attached. May be applied to structures with EncapsulatedContent or EncryptedContent.

isDetached :: Encapsulates struct => struct (Encap a) -> Bool

Return True when the encapsulated content is detached.

fromDetached :: (MonadFail m, Encapsulates struct) => b -> struct (Encap a) -> m (struct b)

Recover the original data structure from a detached encapsulation and the external content. The monadic computation fails if the content was attached.

toDetachedCI :: Encapsulates struct => struct ByteString -> (ByteString, ContentInfo)

Transform the data structure into a content info, detaching the encapsulated content. May be applied to structures with EncapsulatedContent or EncryptedContent.

Signed data

type SignatureValue = ByteString

Signature value.

data SignatureAlg

CMS signature algorithms and associated parameters.

Constructors

RSAAnyHash

RSA DigestAlgorithm

RSAPSS PSSParams

DSA DigestAlgorithm

ECDSA DigestAlgorithm

Ed25519

Ed448

type EncapsulatedContent = ByteString

Encapsulated content.

data SignedData content

Signed content information.

Constructors

SignedData

Fields

sdDigestAlgorithms :: [DigestAlgorithm]

Digest algorithms

sdContentType :: ContentType

Inner content type

sdEncapsulatedContent :: content

Encapsulated content

sdCertificates :: [CertificateChoice]

The collection of certificates

sdCRLs :: [RevocationInfoChoice]

The collection of CRLs

sdSignerInfos :: [SignerInfo]

Per-signer information

type ProducerOfSI m = ContentType -> ByteString -> m (Either StoreError (SignerInfo, [CertificateChoice], [RevocationInfoChoice]))

Function able to produce a SignerInfo.

type ConsumerOfSI m = ContentType -> ByteString -> SignerInfo -> [CertificateChoice] -> [RevocationInfoChoice] -> m Bool

Function able to consume a SignerInfo.

signData :: Applicative f => [ProducerOfSI f] -> ContentInfo -> f (Either StoreError (SignedData EncapsulatedContent))

Add a signed-data layer on the specified content info. The content is processed by one or several ProducerOfSI functions to create signer info elements.

verifySignedData :: Monad m => ConsumerOfSI m -> SignedData EncapsulatedContent -> m (Either StoreError ContentInfo)

Verify a signed content info using the specified ConsumerOfSI function. Verification of at least one signer info must be successful in order to return the inner content info.

Signer information

data SignerInfo

Information related to a signer of a SignedData. An element contains the signature material that was produced.

Constructors

SignerInfo

Fields

siSignerId :: SignerIdentifier

Identifier of the signer certificate

siDigestAlgorithm :: DigestAlgorithm

Digest algorithm used for the signature

siSignedAttrs :: [Attribute]

Optional signed attributes

siSignatureAlg :: SignatureAlg

Algorithm used for signature

siSignature :: SignatureValue

The signature value

siUnsignedAttrs :: [Attribute]

Optional unsigned attributes

data SignerIdentifier

Union type related to identification of the signer certificate.

Constructors

SignerIASN IssuerAndSerialNumber
Issuer and Serial Number

SignerSKI ByteString
Subject Key Identifier

data IssuerAndSerialNumber

Identification of a certificate using the issuer DN and serial number.

Constructors

IssuerAndSerialNumber

Fields

iasnIssuer :: DistinguishedName

Distinguished name of the certificate issuer

iasnSerial :: Integer

Issuer-specific certificate serial number

certSigner :: MonadRandom m => SignatureAlg -> PrivKey -> CertificateChain -> Maybe [Attribute] -> [Attribute] -> ProducerOfSI m

Create a signer info with the specified signature algorithm and credentials.

Two lists of optional attributes can be provided. The attributes will be part of message signature when provided in the first list.

When the first list of attributes is provided, even empty list, signature is computed from a digest of the content. When the list of attributes is Nothing, no intermediate digest is used and the signature is computed from the full message.

withPublicKey :: Applicative f => PubKey -> ConsumerOfSI f

Verify that the signature was produced from the specified public key. Ignores all certificates and CRLs contained in the signed data.

withSignerKey :: Applicative f => ConsumerOfSI f

Verify that the signature is valid with one of the X.509 certificates contained in the signed data, but does not validate that the certificates are valid. All transmitted certificates are implicitely trusted and all CRLs are ignored.

withSignerCertificate :: Applicative f => (Maybe DateTime -> CertificateChain -> f Bool) -> ConsumerOfSI f

Verify that the signature is valid with one of the X.509 certificates contained in the signed data, and verify that the signer certificate is valid using the validation function supplied. All CRLs are ignored.

Enveloped data

type EncryptedKey = ByteString

Encrypted key.

data KeyEncryptionParams

Key encryption algorithm with associated parameters (i.e. the underlying encryption algorithm).

Constructors

PWRIKEK ContentEncryptionParams
PWRI-KEK key wrap algorithm

AES128_WRAP
AES-128 key wrap

AES192_WRAP
AES-192 key wrap

AES256_WRAP
AES-256 key wrap

AES128_WRAP_PAD
AES-128 extended key wrap

AES192_WRAP_PAD
AES-192 extended key wrap

AES256_WRAP_PAD
AES-256 extended key wrap

DES_EDE3_WRAP
Triple-DES key wrap

RC2_WRAP Int
RC2 key wrap with effective key length

data KeyTransportParams

Key transport algorithm with associated parameters.

Constructors

RSAES
RSAES-PKCS1

RSAESOAEP OAEPParams
RSAES-OAEP

data KeyAgreementParams

Key agreement algorithm with associated parameters.

Constructors

StdDH KeyAgreementKDF KeyEncryptionParams
1-Pass D-H with Stardard ECDH

CofactorDH KeyAgreementKDF KeyEncryptionParams
1-Pass D-H with Cofactor ECDH

data KeyAgreementKDF

Key derivation function used for key agreement.

Constructors

forall hashAlg.HashAlgorithm hashAlg => KA_X963_KDF (DigestProxy hashAlg)
ANSI-X9.63-KDF key derivation function

forall hashAlg.HashAlgorithm hashAlg => KA_HKDF (DigestProxy hashAlg)
Extract-and-Expand HMAC-based key derivation function

data RecipientInfo

Information for a recipient of an EnvelopedData. An element contains the content-encryption key in encrypted form.

Constructors

KTRI KTRecipientInfo
Recipient using key transport

KARI KARecipientInfo
Recipient using key agreement

KEKRI KEKRecipientInfo
Recipient using key encryption

PasswordRI PasswordRecipientInfo
Recipient using password-based protection

data EnvelopedData content

Enveloped content information.

Constructors

EnvelopedData

Fields

evOriginatorInfo :: OriginatorInfo

Optional information about the originator

evRecipientInfos :: [RecipientInfo]

Information for recipients, allowing to decrypt the content

evContentType :: ContentType

Inner content type

evContentEncryptionParams :: ContentEncryptionParams

Encryption algorithm

evEncryptedContent :: content

Encrypted content info

evUnprotectedAttrs :: [Attribute]

Optional unprotected attributes

type ProducerOfRI m = ContentEncryptionKey -> m (Either StoreError RecipientInfo)

Function able to produce a RecipientInfo.

type ConsumerOfRI m = RecipientInfo -> m (Either StoreError ContentEncryptionKey)

Function able to consume a RecipientInfo.

envelopData :: Applicative f => OriginatorInfo -> ContentEncryptionKey -> ContentEncryptionParams -> [ProducerOfRI f] -> [Attribute] -> ContentInfo -> f (Either StoreError (EnvelopedData EncryptedContent))

Add an enveloped-data layer on the specified content info. The content is encrypted with specified key and algorithm. The key is then processed by one or several ProducerOfRI functions to create recipient info elements.

Some optional attributes can be added but will not be encrypted.

openEnvelopedData :: Monad m => ConsumerOfRI m -> EnvelopedData EncryptedContent -> m (Either StoreError ContentInfo)

Recover an enveloped content info using the specified ConsumerOfRI function.

Key Transport recipients

data KTRecipientInfo

Recipient using key transport.

Constructors

KTRecipientInfo

Fields

ktRid :: RecipientIdentifier

identifier of recipient

ktKeyTransportParams :: KeyTransportParams

key transport algorithm

ktEncryptedKey :: EncryptedKey

encrypted content-encryption key

data RecipientIdentifier

Union type related to identification of the recipient.

Constructors

RecipientIASN IssuerAndSerialNumber
Issuer and Serial Number

RecipientSKI ByteString
Subject Key Identifier

forKeyTransRecipient :: MonadRandom m => SignedCertificate -> KeyTransportParams -> ProducerOfRI m

Generate a Key Transport recipient from a certificate and desired algorithm. The recipient will contain certificate identifier.

This function can be used as parameter to envelopData.

withRecipientKeyTrans :: MonadRandom m => PrivKey -> ConsumerOfRI m

Use a Key Transport recipient, knowing the private key.

This function can be used as parameter to openEnvelopedData.

Key Agreement recipients

data KARecipientInfo

Recipient using key agreement.

Constructors

KARecipientInfo

Fields

kaOriginator :: OriginatorIdentifierOrKey

identifier of orginator or anonymous key

kaUkm :: Maybe UserKeyingMaterial

user keying material

kaKeyAgreementParams :: KeyAgreementParams

key agreement algorithm

kaRecipientEncryptedKeys :: [RecipientEncryptedKey]

encrypted content-encryption key for one or multiple recipients

data OriginatorIdentifierOrKey

Union type related to identification of the originator.

Constructors

OriginatorIASN IssuerAndSerialNumber
Issuer and Serial Number

OriginatorSKI ByteString
Subject Key Identifier

OriginatorPublic OriginatorPublicKey
Anonymous public key

data OriginatorPublicKey

Originator public key used for key-agreement. Contrary to PubKey the domain parameters are not used and may be left empty.

data RecipientEncryptedKey

Encrypted key for a recipient in a key-agreement RI.

Constructors

RecipientEncryptedKey

Fields

rekRid :: KeyAgreeRecipientIdentifier

identifier of recipient

rekEncryptedKey :: EncryptedKey

encrypted content-encryption key

data KeyAgreeRecipientIdentifier

Union type related to identification of a key-agreement recipient.

Constructors

KeyAgreeRecipientIASN IssuerAndSerialNumber
Issuer and Serial Number

KeyAgreeRecipientKI KeyIdentifier
Key identifier

type UserKeyingMaterial = ByteString

User keying material.

forKeyAgreeRecipient :: MonadRandom m => SignedCertificate -> KeyAgreementParams -> ProducerOfRI m

Generate a Key Agreement recipient from a certificate and desired algorithm. The recipient info will contain an ephemeral public key.

This function can be used as parameter to envelopData.

To avoid decreasing the security strength, Key Encryption parameters should use a key size equal or greater than the content encryption key.

withRecipientKeyAgree :: MonadRandom m => PrivKey -> SignedCertificate -> ConsumerOfRI m

Use a Key Agreement recipient, knowing the recipient private key. The recipient certificate is also required to locate which encrypted key to use.

This function can be used as parameter to openEnvelopedData.

Key Encryption Key recipients

data KEKRecipientInfo

Recipient using key encryption.

Constructors

KEKRecipientInfo

Fields

kekId :: KeyIdentifier

identifier of key encryption key

kekKeyEncryptionParams :: KeyEncryptionParams

key encryption algorithm

kekEncryptedKey :: EncryptedKey

encrypted content-encryption key

data KeyIdentifier

Key identifier and optional attributes.

Constructors

KeyIdentifier

Fields

keyIdentifier :: ByteString

identifier of the key

keyDate :: Maybe DateTime

optional timestamp

keyOther :: Maybe OtherKeyAttribute

optional information

data OtherKeyAttribute

Additional information in a KeyIdentifier.

Constructors

OtherKeyAttribute

Fields

keyAttrId :: OID

attribute identifier

keyAttr :: [ASN1]

attribute value

type KeyEncryptionKey = ByteString

Key used for key encryption.

forKeyRecipient :: MonadRandom m => KeyEncryptionKey -> KeyIdentifier -> KeyEncryptionParams -> ProducerOfRI m

Generate a Key Encryption Key recipient from a key encryption key and desired algorithm. The recipient may identify the KEK that was used with the supplied identifier.

This function can be used as parameter to envelopData.

To avoid decreasing the security strength, Key Encryption parameters should use a key size equal or greater than the content encryption key.

withRecipientKey :: Applicative f => KeyEncryptionKey -> ConsumerOfRI f

Use a Key Encryption Key recipient, knowing the key encryption key.

This function can be used as parameter to openEnvelopedData.

Password recipients

data PasswordRecipientInfo

Recipient using password-based protection.

Constructors

PasswordRecipientInfo

Fields

priKeyDerivationFunc :: KeyDerivationFunc

function to derive key

priKeyEncryptionParams :: KeyEncryptionParams

key encryption algorithm

priEncryptedKey :: EncryptedKey

encrypted content-encryption key

forPasswordRecipient :: MonadRandom m => Password -> KeyDerivationFunc -> KeyEncryptionParams -> ProducerOfRI m

Generate a password recipient from a password.

This function can be used as parameter to envelopData.

withRecipientPassword :: Applicative f => Password -> ConsumerOfRI f

Use a password recipient, knowing the password.

This function can be used as parameter to openEnvelopedData.

Digested data

data DigestProxy hashAlg where

CMS digest proxy. Acts like Proxy, i.e. provides a hash algorithm as type parameter. The GADT constructors map to known algorithms.

Constructors

MD2 :: DigestProxy MD2
MD2

MD4 :: DigestProxy MD4
MD4

MD5 :: DigestProxy MD5
MD5

SHA1 :: DigestProxy SHA1
SHA-1

SHA224 :: DigestProxy SHA224
SHA-224

SHA256 :: DigestProxy SHA256
SHA-256

SHA384 :: DigestProxy SHA384
SHA-384

SHA512 :: DigestProxy SHA512
SHA-512

SHAKE128_256 :: DigestProxy (SHAKE128 256)
SHAKE128 (256 bits)

SHAKE256_512 :: DigestProxy (SHAKE256 512)
SHAKE256 (512 bits)

SHAKE128 :: KnownNat n => Proxy n -> DigestProxy (SHAKE128 n)
SHAKE128 (variable size)

SHAKE256 :: KnownNat n => Proxy n -> DigestProxy (SHAKE256 n)
SHAKE256 (variable size)

data DigestAlgorithm

CMS digest algorithm.

Constructors

forall hashAlg.HashAlgorithm hashAlg => DigestAlgorithm (DigestProxy hashAlg)

data DigestedData content

Digested content information.

Constructors

forall hashAlg.HashAlgorithm hashAlg => DigestedData

Fields

ddDigestAlgorithm :: DigestProxy hashAlg

Digest algorithm

ddContentType :: ContentType

Inner content type

ddEncapsulatedContent :: content

Encapsulated content

ddDigest :: Digest hashAlg

Digest value

digestData :: DigestAlgorithm -> ContentInfo -> DigestedData EncapsulatedContent

Add a digested-data layer on the specified content info.

digestVerify :: DigestedData EncapsulatedContent -> Either StoreError ContentInfo

Return the inner content info but only if the digest is valid.

Encrypted data

type ContentEncryptionKey = ByteString

Key used for content encryption.

data ContentEncryptionCipher cipher where

CMS content encryption cipher.

Constructors

DES :: ContentEncryptionCipher DES
DES

DES_EDE2 :: ContentEncryptionCipher DES_EDE2
Triple-DES with 2 keys used in alternative direction

DES_EDE3 :: ContentEncryptionCipher DES_EDE3
Triple-DES with 3 keys used in alternative direction

AES128 :: ContentEncryptionCipher AES128
AES with 128-bit key

AES192 :: ContentEncryptionCipher AES192
AES with 192-bit key

AES256 :: ContentEncryptionCipher AES256
AES with 256-bit key

CAST5 :: ContentEncryptionCipher CAST5
CAST5 (aka CAST-128) with key between 40 and 128 bits

Camellia128 :: ContentEncryptionCipher Camellia128
Camellia with 128-bit key

data ContentEncryptionAlg

Cipher and mode of operation for content encryption.

Constructors

forall c.BlockCipher c => ECB (ContentEncryptionCipher c)
Electronic Codebook

forall c.BlockCipher c => CBC (ContentEncryptionCipher c)
Cipher Block Chaining

CBC_RC2
RC2 in CBC mode

forall c.BlockCipher c => CFB (ContentEncryptionCipher c)
Cipher Feedback

forall c.BlockCipher c => CTR (ContentEncryptionCipher c)
Counter

data ContentEncryptionParams

Content encryption algorithm with associated parameters (i.e. the initialization vector).

A value can be generated with generateEncryptionParams.

type EncryptedContent = ByteString

Encrypted content.

data EncryptedData content

Encrypted content information.

Constructors

EncryptedData

Fields

edContentType :: ContentType

Inner content type

edContentEncryptionParams :: ContentEncryptionParams

Encryption algorithm

edEncryptedContent :: content

Encrypted content info

edUnprotectedAttrs :: [Attribute]

Optional unprotected attributes

generateEncryptionParams :: MonadRandom m => ContentEncryptionAlg -> m ContentEncryptionParams

Generate random parameters for the specified content encryption algorithm.

generateRC2EncryptionParams :: MonadRandom m => Int -> m ContentEncryptionParams

Generate random RC2 parameters with the specified effective key length (in bits).

getContentEncryptionAlg :: ContentEncryptionParams -> ContentEncryptionAlg

Get the content encryption algorithm.

encryptData :: ContentEncryptionKey -> ContentEncryptionParams -> [Attribute] -> ContentInfo -> Either StoreError (EncryptedData EncryptedContent)

Add an encrypted-data layer on the specified content info. The content is encrypted with specified key and algorithm.

Some optional attributes can be added but will not be encrypted.

decryptData :: ContentEncryptionKey -> EncryptedData EncryptedContent -> Either StoreError ContentInfo

Decrypt an encrypted content info using the specified key.

Authenticated data

type AuthenticationKey = ContentEncryptionKey

Key used for authentication.

data MACAlgorithm

Message Authentication Code (MAC) Algorithm.

Constructors

forall hashAlg.HashAlgorithm hashAlg => HMAC (DigestProxy hashAlg)

forall n.KnownNat n => KMAC_SHAKE128 (Proxy n) ByteString

forall n.KnownNat n => KMAC_SHAKE256 (Proxy n) ByteString

type MessageAuthenticationCode = AuthTag

Message authentication code. Equality is time constant.

data AuthenticatedData content

Authenticated content information.

Constructors

AuthenticatedData

Fields

adOriginatorInfo :: OriginatorInfo

Optional information about the originator

adRecipientInfos :: [RecipientInfo]

Information for recipients, allowing to authenticate the content

adMACAlgorithm :: MACAlgorithm

MAC algorithm

adDigestAlgorithm :: Maybe DigestAlgorithm

Optional digest algorithm

adContentType :: ContentType

Inner content type

adEncapsulatedContent :: content

Encapsulated content

adAuthAttrs :: [Attribute]

Optional authenticated attributes

adMAC :: MessageAuthenticationCode

Message authentication code

adUnauthAttrs :: [Attribute]

Optional unauthenticated attributes

generateAuthenticatedData :: Applicative f => OriginatorInfo -> AuthenticationKey -> MACAlgorithm -> Maybe DigestAlgorithm -> [ProducerOfRI f] -> [Attribute] -> [Attribute] -> ContentInfo -> f (Either StoreError (AuthenticatedData EncapsulatedContent))

Add an authenticated-data layer on the specified content info. The content is MACed with the specified key and algorithms. The key is then processed by one or several ProducerOfRI functions to create recipient info elements.

Two lists of optional attributes can be provided. The attributes will be part of message authentication when provided in the first list.

verifyAuthenticatedData :: Monad m => ConsumerOfRI m -> AuthenticatedData EncapsulatedContent -> m (Either StoreError ContentInfo)

Verify the integrity of an authenticated content info using the specified ConsumerOfRI function. The inner content info is returned only if the MAC could be verified.

Authenticated-enveloped data

data AuthContentEncryptionAlg

Cipher and mode of operation for authenticated-content encryption.

Constructors

AUTH_ENC_128
authEnc with 128-bit key

AUTH_ENC_256
authEnc with 256-bit key

CHACHA20_POLY1305
ChaCha20-Poly1305 Authenticated Encryption

forall c.BlockCipher c => CCM (ContentEncryptionCipher c)
Counter with CBC-MAC

forall c.BlockCipher c => GCM (ContentEncryptionCipher c)
Galois Counter Mode

data AuthContentEncryptionParams

Authenticated-content encryption algorithm with associated parameters (i.e. the nonce).

A value can be generated with functions generateAuthEnc128Params, generateAuthEnc256Params, generateChaChaPoly1305Params, generateCCMParams and generateGCMParams.

data AuthEnvelopedData content

Authenticated-enveloped content information.

Constructors

AuthEnvelopedData

Fields

aeOriginatorInfo :: OriginatorInfo

Optional information about the originator

aeRecipientInfos :: [RecipientInfo]

Information for recipients, allowing to decrypt the content

aeContentType :: ContentType

Inner content type

aeContentEncryptionParams :: ASN1ObjectExact AuthContentEncryptionParams

Encryption algorithm

aeEncryptedContent :: content

Encrypted content info

aeAuthAttrs :: [Attribute]

Optional authenticated attributes

aeMAC :: MessageAuthenticationCode

Message authentication code

aeUnauthAttrs :: [Attribute]

Optional unauthenticated attributes

generateAuthEnc128Params :: MonadRandom m => PBKDF2_PRF -> ContentEncryptionAlg -> MACAlgorithm -> m AuthContentEncryptionParams

Generate random AUTH_ENC_128 parameters with the specified algorithms.

generateAuthEnc256Params :: MonadRandom m => PBKDF2_PRF -> ContentEncryptionAlg -> MACAlgorithm -> m AuthContentEncryptionParams

Generate random AUTH_ENC_256 parameters with the specified algorithms.

generateChaChaPoly1305Params :: MonadRandom m => m AuthContentEncryptionParams

Generate random CHACHA20_POLY1305 parameters.

generateCCMParams :: (MonadRandom m, BlockCipher c) => ContentEncryptionCipher c -> CCM_M -> CCM_L -> m AuthContentEncryptionParams

Generate random CCM parameters for the specified cipher.

generateGCMParams :: (MonadRandom m, BlockCipher c) => ContentEncryptionCipher c -> Int -> m AuthContentEncryptionParams

Generate random GCM parameters for the specified cipher.

authEnvelopData :: Applicative f => OriginatorInfo -> ContentEncryptionKey -> AuthContentEncryptionParams -> [ProducerOfRI f] -> [Attribute] -> [Attribute] -> ContentInfo -> f (Either StoreError (AuthEnvelopedData EncryptedContent))

Add an authenticated-enveloped-data layer on the specified content info. The content is encrypted with specified key and algorithm. The key is then processed by one or several ProducerOfRI functions to create recipient info elements.

Some attributes can be added but will not be encrypted. The attributes will be part of message authentication when provided in the first list.

openAuthEnvelopedData :: Monad m => ConsumerOfRI m -> AuthEnvelopedData EncryptedContent -> m (Either StoreError ContentInfo)

Recover an authenticated-enveloped content info using the specified ConsumerOfRI function.

Key derivation

type Salt = ByteString

Salt value used for key derivation.

generateSalt :: MonadRandom m => Int -> m Salt

Generate a random salt with the specified length in bytes. To be most effective, the length should be at least 8 bytes.

data KeyDerivationFunc

Key derivation algorithm and associated parameters.

Constructors

PBKDF2
Key derivation with PBKDF2

Fields

pbkdf2Salt :: Salt

Salt value

pbkdf2IterationCount :: Int

Iteration count

pbkdf2KeyLength :: Maybe Int

Optional key length

pbkdf2Prf :: PBKDF2_PRF

Pseudorandom function

Scrypt
Key derivation with Scrypt

Fields

scryptSalt :: Salt

Salt value

scryptN :: Word64

N value

scryptR :: Int

R value

scryptP :: Int

P value

scryptKeyLength :: Maybe Int

Optional key length

data PBKDF2_PRF

Pseudorandom function used for PBKDF2.

Constructors

PBKDF2_SHA1
hmacWithSHA1

PBKDF2_SHA256
hmacWithSHA256

PBKDF2_SHA512
hmacWithSHA512

Secret-key algorithms

class HasKeySize params where

Algorithms that are based on a secret key. This includes ciphers but also MAC algorithms.

Methods

getKeySizeSpecifier :: params -> KeySizeSpecifier

Get a specification of the key sizes allowed by the algorithm.

generateKey :: (HasKeySize params, MonadRandom m, ByteArray key) => params -> m key

Generate a random key suitable for the specified algorithm. This uses the maximum size allowed by the parameters.

RSA padding modes

newtype MaskGenerationFunc

Mask Generation Functions (MGF) and associated parameters.

Constructors

MGF1 DigestAlgorithm

data OAEPParams

Encryption parameters for RSAES-OAEP.

Constructors

OAEPParams

Fields

oaepHashAlgorithm :: DigestAlgorithm

Hash function

oaepMaskGenAlgorithm :: MaskGenerationFunc

Mask generation function

data PSSParams

Signature parameters for RSASSA-PSS.

Constructors

PSSParams

Fields

pssHashAlgorithm :: DigestAlgorithm

Hash function

pssMaskGenAlgorithm :: MaskGenerationFunc

Mask generation function

pssSaltLength :: Int

Length of the salt in bytes

CMS attributes

data Attribute

An attribute extending the parent structure with arbitrary data.

Constructors

Attribute

Fields

attrType :: OID

Attribute type

attrValues :: [ASN1]

Attribute values

findAttribute :: OID -> [Attribute] -> Maybe [ASN1]

Return the values for the first attribute with the specified type.

setAttribute :: OID -> [ASN1] -> [Attribute] -> [Attribute]

Add or replace an attribute in a list of attributes.

filterAttributes :: (OID -> Bool) -> [Attribute] -> [Attribute]

Filter a list of attributes based on a predicate applied to attribute type.

CMS standard attributes

getSigningTimeAttr :: [Attribute] -> Maybe DateTime

Return the value of the signingTime attribute.

setSigningTimeAttr :: DateTime -> [Attribute] -> [Attribute]

Add or replace the signingTime attribute in a list of attributes.

setSigningTimeAttrCurrent :: MonadIO m => [Attribute] -> m [Attribute]

Add or replace the signingTime attribute in a list of attributes with the current time. This is equivalent to calling setSigningTimeAttr with the result of dateCurrent.

Originator information

data OriginatorInfo

Information about the originator of the content info, to be used when a key management algorithm requires this information.

Constructors

OriginatorInfo

Fields

originatorCerts :: [CertificateChoice]

The collection of certificates

originatorCRLs :: [RevocationInfoChoice]

The collection of CRLs

data CertificateChoice

Union type related to certificate formats.

Constructors

CertificateCertificate SignedCertificate
X.509 certificate

CertificateOther OtherCertificateFormat
Other format

data OtherCertificateFormat

Certificate information in a format not supported natively.

Constructors

OtherCertificateFormat

Fields

otherCertFormat :: OID

Format identifier

otherCertValues :: [ASN1]

ASN.1 values using this format

data RevocationInfoChoice

Union type related to revocation info formats.

Constructors

RevocationInfoCRL SignedCRL
A CRL, ARL, Delta CRL, or an ACRL

RevocationInfoOther OtherRevocationInfoFormat
Other format

data OtherRevocationInfoFormat

Revocation information in a format not supported natively.

Constructors

OtherRevocationInfoFormat

Fields

otherRevInfoFormat :: OID

Format identifier

otherRevInfoValues :: [ASN1]

ASN.1 values using this format

ASN.1 representation

data ASN1ObjectExact a

An ASN.1 object associated with the raw data it was parsed from.