Crypto.Store.KeyWrap.AES

Description

AES Key Wrap (RFC 3394) and Extended Key Wrap (RFC 5649)

Should be used with a cipher from module Crypto.Cipher.AES.

Documentation

wrap :: (BlockCipher aes, ByteArray ba) => aes -> ba -> Either StoreError ba

Wrap a key with the specified AES cipher.

unwrap :: (BlockCipher aes, ByteArray ba) => aes -> ba -> Either StoreError ba

Unwrap an encrypted key with the specified AES cipher.

wrapPad :: (BlockCipher aes, ByteArray ba) => aes -> ba -> Either StoreError ba

Pad and wrap a key with the specified AES cipher.

unwrapPad :: (BlockCipher aes, ByteArray ba) => aes -> ba -> Either StoreError ba

Unwrap and unpad an encrypted key with the specified AES cipher.