Crypto.Store.X509

Description

Public keys, certificates and CRLs.

Presents an API similar to Data.X509.Memory and Data.X509.File but provides support for public-key files and allows to write objects.

Functions related to private keys are available from Crypto.Store.PKCS8.

Documentation

class (ASN1Object a, Eq a, Show a) => SignedObject a

Class of signed objects convertible to PEM.

Minimal complete definition

signedObjectName, otherObjectNames

Public keys

readPubKeyFile :: FilePath -> IO [PubKey]

Read public keys from a PEM file.

readPubKeyFileFromMemory :: ByteString -> [PubKey]

Read public keys from a bytearray in PEM format.

pemToPubKey :: [Maybe PubKey] -> PEM -> [Maybe PubKey]

Read a public key from a PEM element and add it to the accumulator list.

writePubKeyFile :: FilePath -> [PubKey] -> IO ()

Write public keys to a PEM file.

writePubKeyFileToMemory :: [PubKey] -> ByteString

Write public keys to a bytearray in PEM format.

pubKeyToPEM :: PubKey -> PEM

Generate a PEM for a public key.

Signed objects

readSignedObject :: SignedObject a => FilePath -> IO [SignedExact a]

Read signed objects from a PEM file (only one type at a time).

readSignedObjectFromMemory :: SignedObject a => ByteString -> [SignedExact a]

Read signed objects from a bytearray in PEM format (only one type at a time).

writeSignedObject :: SignedObject a => FilePath -> [SignedExact a] -> IO ()

Write signed objects to a PEM file.

writeSignedObjectToMemory :: SignedObject a => [SignedExact a] -> ByteString

Write signed objects to a bytearray in PEM format.

Reading and writing PEM files

readPEMs :: FilePath -> IO [PEM]

Read a PEM file from disk.

writePEMs :: FilePath -> [PEM] -> IO ()

Write a PEM file to disk.